+++ packages/examples/complete-demo/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "complete-demo",
+  "name": "@opuscapita/fsm-workflow-demo",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
@@ -7,13 +7,17 @@
     "start": "cross-env PORT=8082 node ./node_modules/babel-cli/bin/babel-node.js ./src/server/index.js --presets=env",
     "start2": "webpack-dev-server --config src/server/webpack.config.js",
     "build": "webpack --config src/server/webpack.config.js",
-    "starthttp": "http-server"
+    "add-dependency": "cd ../../../ && npm run add-dependency --scope=@opuscapita/fsm-workflow-demo --"
   },
   "dependencies": {
+    "@opuscapita/fsm-workflow-core": "1.0.0",
     "@opuscapita/fsm-workflow-editor": "1.0.0",
     "@opuscapita/i18n": "1.1.1",
+    "@opuscapita/react-navigation": "0.1.2",
     "connect-timeout": "1.9.0",
     "express": "4.16.2",
+    "json-schema-faker": "0.4.7",
+    "lodash": "^4.17.5",
     "prop-types": "15.6.0",
     "react-bootstrap": "0.32.1"
   },
