+++ packages/examples/complete-demo/src/client/components/utils.js
@@ -1,0 +1,6 @@
+export const mapFuncsToParamsSchema = funcsObj => Object.keys(funcsObj).reduce((res, name) => ({
+  ...res,
+  [name]: {
+    paramsSchema: funcsObj[name].paramsSchema
+  }
+}), {});
