+++ packages/examples/complete-demo/src/client/customComponentsRegistry/index.js
@@ -1,0 +1,5 @@
+import FullName from './FullName.react';
+
+export default {
+  fullName: FullName
+}
