+++ packages/examples/complete-demo/src/client/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './src/App.react';
+import App from './components/App.react';
 
 ReactDOM.render(<App />, document.getElementById('main'));
