--- packages/examples/complete-demo/src/client/src/demo-data/actions/index.js
@@ -1,9 +1,0 @@
-import testAction from './testAction';
-import sendMail from './sendMail';
-import updateProcessedBy from './updateProcessedBy';
-
-export default {
-  testAction,
-  sendMail,
-  updateProcessedBy
-}
