--- packages/examples/complete-demo/src/client/src/demo-data/actions/updateProcessedBy.js
@@ -1,15 +1,0 @@
-const action = args => {
-  console.log('action "updateProcessedBy" received args: ', args)
-}
-
-action.paramsSchema = {
-  "type": "object",
-  "properties": {
-    processedByFieldName: {
-      "type": "string"
-    }
-  },
-  "required": ["processedByFieldName"]
-}
-
-export default action;
