--- packages/examples/complete-demo/src/client/src/demo-data/components/index.js
@@ -1,5 +1,0 @@
-import FullName from './FullName.react';
-
-export default {
-  fullName: FullName
-}
