--- packages/examples/complete-demo/src/client/src/demo-data/conditions/index.js
@@ -1,5 +1,0 @@
-import userHasRoles from './userHasRoles';
-
-export default {
-  userHasRoles
-}
