--- packages/examples/complete-demo/src/client/src/demo-data/conditions/userHasRoles.js
@@ -1,18 +1,0 @@
-const condition = args => {
-  console.log('condition "userHasRoles" received args: ', args)
-}
-
-condition.paramsSchema = {
-  type: "object",
-  properties: {
-    restrictedRoles: {
-      type: "array",
-      items: {
-        type: "string",
-        enum: ["REV", "BOSS"]
-      }
-    }
-  }
-}
-
-export default condition;
