--- packages/examples/complete-demo/src/client/src/demo-data/workflow-schema.json
@@ -1,237 +1,0 @@
-{
-  "name": "InvoiceApproval",
-  "initialState": "inspectionRequired",
-  "finalStates": [
-    "approved"
-  ],
-  "transitions": [
-    {
-      "from": "inspectionRequired",
-      "to": "approvalRequired",
-      "event": "inspect",
-      "guards": [
-        {
-          "expression": "invoice[\"netAmount\"] > 100"
-        },
-        {
-          "name": "userHasRoles",
-          "params": [
-            {
-              "name": "restrictedRoles",
-              "value": [
-                "REV"
-              ]
-            }
-          ]
-        }
-      ],
-      "actions": [
-        {
-          "name": "testAction",
-          "params": [
-            {
-              "name": "dinnerMenu",
-              "value": [
-                "Steak",
-                "Mashrooms"
-              ]
-            },
-            {
-              "name": "fullName",
-              "value": "John Smith"
-            }
-          ]
-        },
-        {
-          "name": "sendMail",
-          "params": [
-            {
-              "name": "fromAddress",
-              "value": "support@client.com"
-            },
-            {
-              "name": "greeting",
-              "value": "Mr. Twister"
-            }
-          ]
-        },
-        {
-          "name": "updateProcessedBy",
-          "params": [
-            {
-              "name": "someArray",
-              "value": [
-                "one",
-                "two"
-              ]
-            },
-            {
-              "name": "processedByFieldName",
-              "value": "customerId"
-            }
-          ]
-        }
-      ]
-    },
-    {
-      "from": "inspectionRequired",
-      "to": "approvalRequired",
-      "event": "automatic-inspect"
-    },
-    {
-      "from": "inspectionRequired",
-      "to": "inspClrRequired",
-      "event": "sendToClarification"
-    },
-    {
-      "from": "inspClrRequired",
-      "to": "inspectionRequired",
-      "event": "clarifyForInspection"
-    },
-    {
-      "from": "inspectionRequired",
-      "to": "inspectionRejected",
-      "event": "rejectInspection"
-    },
-    {
-      "from": "inspClrRequired",
-      "to": "inspectionRejected",
-      "event": "rejectInspection"
-    },
-    {
-      "from": "inspectionRejected",
-      "to": "inspectionRequired",
-      "event": "cancelRejection"
-    },
-    {
-      "from": "approvalRequired",
-      "to": "inspectionRequired",
-      "event": "cancelInspection"
-    },
-    {
-      "from": "approvalRequired",
-      "to": "inspectionRejected",
-      "event": "rejectInspection"
-    },
-    {
-      "from": "approvalRequired",
-      "to": "approved",
-      "event": "approve"
-    },
-    {
-      "from": "approvalRequired",
-      "to": "appClrRequired",
-      "event": "sendToClarification"
-    },
-    {
-      "from": "approvalRequired",
-      "to": "inspClrRequired",
-      "event": "sendToClarification"
-    },
-    {
-      "from": "approvalRequired",
-      "to": "approvalRejected",
-      "event": "rejectApproval"
-    },
-    {
-      "from": "appClrRequired",
-      "to": "approvalRequired",
-      "event": "clarifyForApproval"
-    },
-    {
-      "from": "appClrRequired",
-      "to": "approvalRejected",
-      "event": "rejectApproval"
-    },
-    {
-      "from": "approved",
-      "to": "approvalRequired",
-      "event": "cancelApproval"
-    },
-    {
-      "from": "approved",
-      "to": "approvalRejected",
-      "event": "rejectApproval"
-    },
-    {
-      "from": "approved",
-      "to": "appClrRequired",
-      "event": "sendToClarification"
-    },
-    {
-      "from": "approvalRejected",
-      "to": "approvalRequired",
-      "event": "cancelRejection"
-    }
-  ],
-  "states": [
-    {
-      "name": "inspectionRequired",
-      "description": "Inspection Required"
-    },
-    {
-      "name": "approvalRequired"
-    },
-    {
-      "name": "inspClrRequired"
-    },
-    {
-      "name": "inspectionRejected"
-    },
-    {
-      "name": "approved"
-    },
-    {
-      "name": "appClrRequired"
-    },
-    {
-      "name": "approvalRejected"
-    }
-  ],
-  "objectConfiguration": {
-    "stateFieldName": "status",
-    "alias": "invoice",
-    "schema": {
-      "type": "object",
-      "properties": {
-        "invoiceNo": {
-          "type": "string"
-        },
-        "customerId": {
-          "type": "string"
-        },
-        "supplierId": {
-          "type": "string"
-        },
-        "netAmount": {
-          "type": "integer"
-        },
-        "grossAmount": {
-          "type": "integer"
-        },
-        "vatAmount": {
-          "type": "integer"
-        },
-        "currencyId": {
-          "type": "string"
-        },
-        "status": {
-          "type": "string"
-        }
-      },
-      "required": [
-        "invoiceNo"
-      ]
-    },
-    "example": {
-      "invoiceNo": "1111",
-      "customerId": "wefwefewfew",
-      "supplierId": "33333",
-      "netAmount": 1000,
-      "grossAmount": 1200,
-      "vatAmount": 200,
-      "currencyId": "EUR",
-      "status": "reviewRequired"
-    }
-  }
-}
