--- packages/examples/complete-demo/src/client/src/utils.js
@@ -1,6 +1,0 @@
-export const mapFuncsToParamsSchema = funcsObj => Object.keys(funcsObj).reduce((res, name) => ({
-  ...res,
-  [name]: {
-    paramsSchema: funcsObj[name].paramsSchema
-  }
-}), {});
