+++ packages/examples/complete-demo/src/server/data/actions/index.js
@@ -1,0 +1,9 @@
+import testAction from './testAction';
+import sendMail from './sendMail';
+import updateProcessedBy from './updateProcessedBy';
+
+export default {
+  testAction,
+  sendMail,
+  updateProcessedBy
+}
