+++ packages/examples/complete-demo/src/server/data/actions/sendMail.js
@@ -1,0 +1,39 @@
+const action = args => {
+  console.log('action "sendMail" received args: ', args)
+}
+
+action.paramsSchema = {
+  "type": "object",
+  "properties": {
+    fromAddress: {
+      "type": "string"
+    },
+    greeting: {
+      "type": "string"
+    },
+    sendCopy: {
+      "type": "boolean"
+    },
+    maxRetries: {
+      "type": "integer"
+    },
+    interest: {
+      "type": "number"
+    },
+    language: {
+      "type": "string",
+      "enum": ["en", "de", "fi", "ru", "sv", "no"]
+    },
+    priority: {
+      "type": "integer",
+      "enum": [0, 1, 2, 3, 4, 5]
+    },
+    expiryDate: {
+      "type": "string",
+      "format": "date"
+    }
+  },
+  "required": ["fromAddress", "greeting"]
+}
+
+export default action;
