+++ packages/examples/complete-demo/src/server/data/actions/updateProcessedBy.js
@@ -1,0 +1,15 @@
+const action = args => {
+  console.log('action "updateProcessedBy" received args: ', args)
+}
+
+action.paramsSchema = {
+  "type": "object",
+  "properties": {
+    processedByFieldName: {
+      "type": "string"
+    }
+  },
+  "required": ["processedByFieldName"]
+}
+
+export default action;
