+++ packages/examples/complete-demo/src/server/data/conditions/index.js
@@ -1,0 +1,5 @@
+import userHasRoles from './userHasRoles';
+
+export default {
+  userHasRoles
+}
