+++ packages/examples/complete-demo/src/server/data/conditions/userHasRoles.js
@@ -1,0 +1,18 @@
+const condition = args => {
+  console.log('condition "userHasRoles" received args: ', args)
+}
+
+condition.paramsSchema = {
+  type: "object",
+  properties: {
+    restrictedRoles: {
+      type: "array",
+      items: {
+        type: "string",
+        enum: ["REV", "BOSS"]
+      }
+    }
+  }
+}
+
+export default condition;
