+++ packages/examples/complete-demo/www/index.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Workflow Demo App</title>
+  <link rel="stylesheet" href="https://opuscapita.github.io/styles/index.css">
+
+</head>
+
+<body style="padding-bottom: 0; overflow-x: hidden;">
+  <div id="main"></div>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/react/15.6.2/react.js" type="text/javascript" ></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/15.6.2/react-dom.js" type="text/javascript" ></script>
+  <script type="text/javascript" src="bundle.js"></script>
+</body>
+</html>
